Using correct math to handle collisions and responses.

+++ utils.c
@@ -40,8 +40,8 @@ void sleep_ticks(const uint64_t ticks)
 }
 
 /* return a random value between [min,max] */
-int get_random(int min, int max)
+float get_random(float min, float max)
 {
-	int range = ( max - min ) + 1;
+	float range = ( max - min ) + 1;
 	return ((rand() / (RAND_MAX + 1.0)) * range) + min;
 }
